Merge polling logic
There were two types of polling going on, polling for authentication
after the user presses the bridge button, and polling for changes on an
API endpoint. Both looked very similar, except the bridge polling used
POST instead of GET, and sent some data, too. With a few more
configuration options in the `poll` module, the bridge auth polling is
now dependent on the `poll` module, reducing the level of duplication
and making the polling logic more capable as a whole. Whew! What a
commit message.