Fix webpack HMR
Not all the fields required for webpack hot module replacement were
correct, and part of the webpack build was failing. Since this is a more
difficult bug to reproduce, there aren't any automated tests for it,
leaving it undiscovered. It seems the regression was introduced in the
commit hash "de689c4a08".