Get initial hue state
Detects changes more accurately by establishing an initial baseline to
compare against. As a consequence, polling is initiated asynchronously,
only after the initial state has been retrieved.
What did this commit fix? If every change is going to be persisted to a
database, we need to know what constitutes a change. When the server
initially starts, it has no state, so is the initial fetch a change that
should be persisted? This commit says no.