Assume all image deps are prebuilt
Building the docker image is so sloowwwww! Here are the main
performance culprits:
- Installing dependencies
- Compiling the source code

Instead of installing and compiling every time a docker image is built,
for reasons of maintaining sanity, I've simply assumed all dependencies
are installed in node_modules and the dist/ folder is already built. Not
exactly intuitive, but until I find a better solution, I simply can't
(won't) wait for those installs. Copying is faster.